Event manager addition
Added an event manager so that listeners will be in a specific order when called.

+++ BrackeysGamejamFinal/Assets/Scripts/Data Management/Serialization/Game Elements Data/PlayerData.cs
@@ -18,7 +18,7 @@ public class PlayerData
     public float armor;
     public float maxArmor;
     public WeaknessType weakness;
-    public float weaknessFactor;
+    public int weaknessFactor;
     public int fireAttack;
     public int waterAttack;
     public int windAttack;
